requirements.txt
networkx
async-lichess-sdk
numpy
pygraphviz